request.sh better caching of bugs and results
To prepare the data can take several seconds,
but bootp requests usually time out within a second.
Hence the results are cached for 1 minute,
such that the result can be reused in the retried requests.

But bootp is single threaded, so only processes one boot request at a time.
This is a deliberate design choice to keep things as simple as it can get.

If many requests arrive at the same time the processing can
pile up, such that it can take a long time until the requests
are processed.  Requests are not discarded, so they pile up.

Things get worse when there are bugs, as if request processing fails,
it would be repeated over and over again.  Like a DoS attack.

Hence bugs are now cached for 12 seconds after the last buggy request.
Then the bugs are re-evaluated alltogether.

Simlar is done for successful requests, but here the timeout is 24s.

This might not be ideal and still needs investigation if it
really works in congestion situations, but hopefully
the DCHP is able to pull itself out of the swamp this way.